requirements.txt
openai
python-dotenv
feedparser
requests
tweepy
gradio
pydantic
rich
beautifulsoup4
requests
feedparser